+++ internal/core/models/reply_strategy.go
@@ -6,9 +6,9 @@ import "time"
 type ReplyStrategy string
 
 const (
-	StrategyNeverReply ReplyStrategy = "never_reply" // 完全不回复（默认）
+	StrategyNeverReply ReplyStrategy = "never_reply" // 完全不回复
 	StrategyAtOnly     ReplyStrategy = "at_only"     // 仅@我时回复
-	StrategyAlways     ReplyStrategy = "always"      // 完全回复
+	StrategyAlways     ReplyStrategy = "always"      // 完全回复（默认）
 	StrategyPluginOnly ReplyStrategy = "plugin_only" // 仅 Plugin
 	StrategyRelevance  ReplyStrategy = "relevance"   // 按相关性回复
 )
@@ -16,7 +16,7 @@ const (
 // ReplyStrategyConfig 系统回复策略配置（单例，DB 中只保留一行）。
 type ReplyStrategyConfig struct {
 	ID                 string        `gorm:"primaryKey;type:uuid" json:"id"`
-	Strategy           ReplyStrategy `gorm:"not null;default:'never_reply'" json:"strategy"`
+	Strategy           ReplyStrategy `gorm:"not null;default:'always'" json:"strategy"`
 	RelevanceThreshold float64       `gorm:"default:0.5" json:"relevance_threshold"`
 	CreatedAt          time.Time     `json:"created_at"`
 	UpdatedAt          time.Time     `json:"updated_at"`
